Allow access to tmp directories
We run Tor Browser in a Flatpak now, which doesn't allow access to the
host's /tmp and /var/tmp. The Flatpak runtime also unsets the TMPDIR
environment variable, so we can't easily make Tor Browser use a custom
TMPDIR in the Flatpak. Therefore we allow it access to /tmp inside the
Flatpak.